code cleanup * make indents consitent * restructure the code tree
