+++ cloudbrain/modules/sources/openbci.py
@@ -7,10 +7,7 @@ from cloudbrain.modules.interface import ModuleInterface
 _LOGGER = logging.getLogger(__name__)
 
 class OpenBCISource(ModuleInterface):
-  def __init__(self, subscribers, publishers,
-               port='/dev/tty.usbserial-DN0095VT',
-               baud=115200,
-               filter_data=True):
+  def __init__(self, subscribers, publishers, port, baud, filter_data):
 
     super(OpenBCISource, self).__init__(subscribers, publishers)
     self.connector = OpenBCIConnector(port=port, baud=baud,
@@ -21,12 +18,13 @@ class OpenBCISource(ModuleInterface):
 
 
   def start(self):
-    # Callback functions to handle the sample for that metric. 
+
+    # Callback functions to handle the sample for that metric.
     # Each metric has a specific number of channels.
     callback_functions = {}
 
     for publisher in self.publishers:
-      metrics_to_num_channels = publisher.get_metrics_to_num_channels()
+      metrics_to_num_channels = publisher.metrics_to_num_channels()
       logging.debug("Metrics to channels mapping: %s" % metrics_to_num_channels)
       for (metric_name, num_channels) in metrics_to_num_channels.items():
         if metric_name not in callback_functions:
@@ -52,7 +50,7 @@ class OpenBCISource(ModuleInterface):
       for i in range(num_channels):
         channel_value = "%.4f" % (
           sample.channel_data[i] * 10 ** 9)  # Nano volts
-        message["channel_%s" % i] = channel_value
+        message["channel_%s" % i] = float(channel_value)
         message['timestamp'] = int(time.time() * 1000000)  # micro seconds
 
       for publisher in self.publishers:
